Moved authenticate route from routes.ts to authenticate.ts
